fix: Moved NewsFragment
