+++ app/src/main/res/layout/fragment_news.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".spacereporter.ui.fragments.NewsFragment"
+    tools:context=".spacereporter.ui.news.NewsFragment"
     android:scrollbars="vertical">
 
     <LinearLayout
